Code formatting

+++ Backend/routes/customers.js
@@ -7,7 +7,10 @@ const {
   getCustomerFromDb,
   updateCustomerInDb,
 } = require("../services/customerServices");
-const {requireLoggin, requirePermission} = require('../middleware/authMiddleware')
+const {
+  requireLoggin,
+  requirePermission,
+} = require("../middleware/authMiddleware");
 
 // GET endpoint to retrieve all customers
 router.get("/", requireLoggin, async (req, res, next) => {
@@ -21,14 +24,19 @@ router.get("/", requireLoggin, async (req, res, next) => {
 });
 
 // POST endpoint to add a new customer
-router.post("/", requireLoggin, requirePermission('admin'), async (req, res) => {
-  try {
-    const newCustomer = await addCustomerToDb(req.body);
-    res.status(201).json({ result: newCustomer });
-  } catch (err) {
-    res.status(500).send("Server error");
-  }
-});
+router.post(
+  "/",
+  requireLoggin,
+  requirePermission("admin"),
+  async (req, res) => {
+    try {
+      const newCustomer = await addCustomerToDb(req.body);
+      res.status(201).json({ result: newCustomer });
+    } catch (err) {
+      res.status(500).send("Server error");
+    }
+  },
+);
 
 // GET endpoint to retrieve a specific customer by id
 router.get("/:customerId", requireLoggin, async (req, res) => {
@@ -46,19 +54,24 @@ router.get("/:customerId", requireLoggin, async (req, res) => {
 });
 
 // DELETE endpoint to delete a specific customer by id
-router.delete("/:customerId", requireLoggin, requirePermission('admin'), async (req, res) => {
-  const customerId = req.params.customerId;
-  try {
-    const result = await deleteCustomerFromDb(customerId);
-    if (result.success) {
-      res.status(204).send();
-    } else {
-      res.status(404).send("Customer not found");
+router.delete(
+  "/:customerId",
+  requireLoggin,
+  requirePermission("admin"),
+  async (req, res) => {
+    const customerId = req.params.customerId;
+    try {
+      const result = await deleteCustomerFromDb(customerId);
+      if (result.success) {
+        res.status(204).send();
+      } else {
+        res.status(404).send("Customer not found");
+      }
+    } catch (error) {
+      res.status(500).send("Server error");
     }
-  } catch (error) {
-    res.status(500).send("Server error");
-  }
-});
+  },
+);
 
 // PUT endpoint to update a specific customer
 router.put("/:customerId", requireLoggin, async (req, res) => {
